+++ autotarcompress/commands/decrypt.py
@@ -10,9 +10,9 @@ import os
 import re
 import subprocess
 
-from src.commands.command import Command
-from src.config import BackupConfig
-from src.security import ContextManager
+from autotarcompress.commands.command import Command
+from autotarcompress.config import BackupConfig
+from autotarcompress.security import ContextManager
 
 
 class DecryptCommand(Command):
