+++ autotarcompress/config.py
