--- src/commands/__init__.py
@@ -1,22 +1,0 @@
-"""Command pattern implementations for backup operations.
-
-This module aggregates all command classes for easy importing.
-"""
-
-from src.commands.command import Command
-from src.commands.backup import BackupCommand
-from src.commands.cleanup import CleanupCommand
-from src.commands.decrypt import DecryptCommand
-from src.commands.encrypt import EncryptCommand
-from src.commands.extract import ExtractCommand
-from src.commands.info import InfoCommand
-
-__all__ = [
-    "Command",
-    "BackupCommand",
-    "CleanupCommand",
-    "DecryptCommand",
-    "EncryptCommand",
-    "ExtractCommand",
-    "InfoCommand",
-]
