+++ tests/conftest.py
@@ -14,7 +14,7 @@ import pytest
 # Add the parent directory to sys.path so Python can find src
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from src.config import BackupConfig
+from autotarcompress.config import BackupConfig
 
 
 @pytest.fixture
