+++ tests/test_facade.py
@@ -11,8 +11,8 @@ from unittest.mock import patch
 # Add the parent directory to sys.path so Python can find src
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from src.config import BackupConfig
-from src.facade import BackupFacade
+from autotarcompress.config import BackupConfig
+from autotarcompress.facade import BackupFacade
 
 
 class TestBackupFacade:
